tidy bin: move signals into module, name threads
